requirements.txt
keras
numpy
python3-opencv
opencv-python
scipy
tensorflow-cpu
